Backfill Tests
Typically I would leave all the tests under the initial describe block
but this function is doing a ton and I'd rather have the backstage pass
tests outside the describe because it's just so much. If nested
describes were a thing in exunit I might have used that as well.